ENH: MPI: print live messages to console
Until this commit we had to wait for the MPI simulation to complete before any
output would be printed on the console. This commit uses the non-blocking select
call to print stderr and stdout while the simulation is in progress.

With MPIBackend we pass the simulation results back via stderr, which means that
MPI error messages during the simulation. We use base64 padding characters to
separate the data from the MPI messages. A trick with base64 encoding is to use
b'===' to always ensure tha  the padding offset is correct, but also having
something to indicate the boundary before the error message.

Fixes #112