ENH: use Queue in MPIBackend for non-blocking read
Select is not compatible with windows (see #215), so we can
follow the same scheme MNE uses with a thread that reads from
stdin/stdout/stderr and appeands line-by-line to a queue. The
main process can then read from the queue. Implements non-
blocking read and is a bit faster than the old _read_all_bytes
function, even with #287.

Also add cleanup processes on KeyboardInterrupt that kill any
running nrniv processes.

Closes #285 and #215