Update roadmap.rst
I updated the roadmap based on the actual plan of action, and to highlight the things that have already been completed. Can you please review to make sure this all makes sense? Or update and add things accordingly? Does it still make sense to develop two modules in hnn-core:  a "simulator" module and a "analysis" module?  Can we think about separating the "simulator" module into - "cortical network model" containing the cortical model,  and "activation of network model" containing  the function to drive the network with specific patterns of activity? We state in the first section that we have organized the code this way - but have we?