MAINT: use custom version of _enqueue_output
The MNE version of _enqueue_output is not sufficient for our use.
The child process needs to remove newlines and we cleanly shut down
threads in MPIBackend by passing an Event to signal when the sim
has finished.

Fixes up the actions that happen when a simulation is exiting. Functions
that use psutil are updated with docstrings as some may be reused
by hnn.