MAINT: add terminate method to MPIBackend
This method will be called asynchronously by HNN GUI. So we need to
use a lock to protect the variable declaring whether the simulation
should terminate.

Killing nrniv procs are done in two places, and only in hnn-core.
First, when the context manager exits, as this is basically state
state is only valid within the manager. Second, when a terminate()
call is made specifically. The user may desire to run multiple
simulations within a single MPIBackend context manager.