TST: fix MPI oversubscribe test on travis
The os.sched_getaffinity(0) call on Travis (Linux) would return 2,
thus limiting n_procs to 2 for MPIBackend and not allowing the
oversubscribe CI test to run with 3 processors.

Note some clusters set scheduler affinity to limit jobs on shared
resources. This change means that the user will be responsible
for setting n_procs appropriately and not exceeding resource
reservations.