MAINT: use Queue in MPIBackend.terminate()
MPIBackend.terminate is going to be called from another thread than
run_subprocess, so we use Queue as a thread-safe way of passing a
valid process handle that terminate() can use. The presence of a
an object in MPIBackend.proc_queue means that a process is currently
running. It will be removed either when run_subprocess completes
successfully or by terminate. It can also be used by other functions
to tell if there is a pending simulation.