fix: set selected index for figure controls after new figure is generated
When creating a blank figure, the GUI was not automatically opening/selecting the new figure tab. This was because the selected index was only being set for the figure tabs but not the figure control tabs. This made them not synchronous, and when the two tab groups are later linked, it reverted to the un-updated index.