fix: accomodate invert-hist second Axes obj
Since the addition of inverted spike histograms requires the use of two
Axes objects, we need to change a few small things:

1. Add a label to our new, second Axes object upon creation in `viz`
2. Use this label to completely remove the second Axes object when
"Clear axis" button is pressed
3. Since axes of type "input histogram" are now going to have 2 Axes
object, while most others will only have 1, we need to make an exception
in the testing. I have done so in `tests/test_gui.py` along with a small
refactor.

The test refactor consists of moving the content of
`test_gui_adaptive_spectrogram` into just another case of the
`test_gui_visualization` test. Additionally, all tests in that test now
take place in a brand-new "Blank single figure", so that their total
axes number do NOT include those of axes that are not the current
`viz_type` being tested (which was the case before).