refactor network.py::_add_cell_type_bias and docs
When reading the new docstring, I noticed that the type hint for
amplitude did not match the behavior of this function's call from
`Network.add_tonic_bias`. I believe that what happened is, part of the
former body of `add_tonic_bias` was refactored out to its own function,
`_add_cell_type_bias`, and as part of this, the interface was assumed to
be mostly the same. The interface is not the same, however: unlike
`add_tonic_bias`, which has 2 ways of specifying the cell_type <->
amplitude mapping (individual arguments or a single dict), when
`add_tonic_bias` calls `_add_cell_type_bias`, it is always passing a
single mapping of cell_type <-> amplitude. I have updated the type hints
and docstring to account for this.

I also removed the unnecessary checks of required-argument values, and
promoted `cell_type` to a required argument, since previously, if it was
None, the function would have not done anything.

In the future, when we enforce the deprecation warned about in
`add_tonic_bias`, this function should probably be moved back into the
body of `add_tonic_bias`, since it will be a singleton and is simple.