maint: bump versions to dev after v0.4.1 stable
This increases the master version from v0.4.1 (the most recent stable)
to v0.5.0dev0. The addition of the "micro" version specificer (see
https://packaging.python.org/en/latest/specifications/version-specifiers/#version-specifiers
and
https://packaging.python.org/en/latest/discussions/versioning/
) does not appear to cause any problems with the documentation
generation as far as I've been able to test, and enables us to version
bugfix patches in the more common manner.

This also re-locates all the version-programming code in `doc/conf.py`
to the same section. Previously this code was distributed across the
file, but there were also inconsistencies between the string-detection
of development versions for the HTML version switcher versus
Binder. These inconsistencies have been resolved.

Equivalently, in the same code blocks, the development version is
indicated for both the HTML dropdown/version control and binder based on
string-detection of EITHER "rc" or "dev" in the version name. This
allows us to return to using "dev" for non-release-candidate development
versions, but also lets us use "rc" in the future if we so desire.

This also updates the documentation copyright to the current year.