Simplify rename_cell_type (#1045)
* [MAINT]: Simplify

* doc: rename/improve docs of dict string replacer

Old name implied that it only changed keys, but it changes keys and
values.

* maint: move dict-id-replacer to utils

* feat: update dict-id-replacer to handle lists too

Original version did not properly replace identifier strings if they
were inside a list. For example, during testing, an attempted
"L5_pyramidal" -> "L5_pyramidal_test"
rename resulted in everything updating correctly except, for example,
the elements of the `self.external_drives["target_types"]` list
elements. This fixes that and updates the docs for it.

---------

Co-authored-by: Austin E. Soplata <me@asoplata.com>