doc: In-place update to install page
This replaces the current "stable" version of the install page with a
newer version. This was done so that the now-recommended `conda` install
would be featured on the default install page, but without having to
publish a new software release for that change alone.

This `install.html` file was generated based on the `doc/install.md`
version from the latest `master` (62b5d35), but built after having
changed all other documentation to use the stable v0.4.1 version. This
was done in order for the page's version information to be the same as
the rest of stable. (A localhost ignore setting was also needed to be
updated in `doc/conf.py` but that is unnecessary to include this in this
commit, since `gh-pages` ONLY includes the site output files).