[MRG] Introduce `ruff` code-format style (#934)
* Enable use of ruff formatting

This does a few things:
- Adds two new Makefile recipes for "ruff"-style formatting of our code:
`format-check` and `format-overwrite`.
- Adds `format-check` to the list of recipes run when you use the `test`
recipe via `make test`.
- Uses the default ruff formatting style.

More info on `ruff format` specifically can be found here:
https://docs.astral.sh/ruff/formatter/

Description of the two new Makefile recipes:

`format-check` is a non-code-changing recipe that tells the user if the
HNN-Core Python code is correctly formatted according to our preferred
ruff style. Similar to our `lint` recipe in the same file, this will not
change any code in any files, but instead simply outputs logs to the user.

`format-overwrite` is a recipe that **CAN, and probably will, change
your code**. This recipe runs the same formatting check as
`format-check`, but then also changes the formatting of your
code **in-place** (meaning in the file itself) so as to "correct" the
code's formatting. Note that the code changes it makes *should* have no
effect on the actual functioning of your code. Even though it should be
safe, you should ALWAYS run tests after you do a `format-overwrite`,
just to be safe.

* Mega-commit of ruff auto formatting

* Require ruff format check in unix unit tests

* Fix codespell complaint

* Fix setup.py reliance on version single-quotes

* Add formatting descr (and lint/spell) to contrib

* doc: Further refinement of quality control sec

* doc: small addition about formatting ctrl of lines