modl: update GUI netw with new cell spacing
This adds a "regenerated" version of the hierarchical JSON network file
used for the GUI, but including the new changes to cell spacing in #985
. This new network file was created by running the
`regenerate_base_network.py` script from #1068 in #985's context.

There are some negligible changes to floating-point values of `end_pts`,
and the addition of a `gain` key, but those should not cause any
effective changes to data output.

The main reason for doing this change is to include the changes to the
`pos_dict` locations of the celltypes as described in #985. However,
these are also not expected to cause changes to *most* of HNN's usage,
including dipole or spike data output. However, they are expected to
produce small and probably insignificant changes in LFP/CSD output.