test: remove linkchecks of specific github iss/PRs
This adds a new ignore regex to Sphinx's linkcheck, such that it will no
longer check the the high number of links that point to Github issues
and PRs in our documentation.

This group of links informs the vast majority of the links in our
documentation, and also the vast majority of the time that linkcheck
takes. These are also a frequent cause of linkcheck failures, due to
timeouts and other Github hosting issues that are not tied to errors in
our documentation linkage itself.

This also converts the former localhost ignore in linkcheck to one that
ignore references to localhost on all ports.