doc/test: add optional, expanded "linkcheck-all"
This adds a new, optional type of linkcheck, called by using the
`doc/Makefile` recipe `linkcheck-all`. This runs linkcheck like the
regular linkcheck recipe, except that it ignores different links. This
is meant to be used whenever a developer wants to test that the links
now ignored in regular linkcheck by #1069 are still valid.

This new recipe only ignores links that:

1. follow any regular expressions used in `doc/conf.py`, or
2. are known to cause errors.

The links which are ignored currently are 3 paper DOI links (which
return 403 Forbidden errors if linkcheck'd), one reference to lines of
code in a specific commit of MNE which returns an error, and the
specific localhost URL for our GUI which also returns an error.