[MRG] Bugfix: Update vecevent mod (#1058)
* Update vecevent.mod

This replaces our `vecevent.mod` file with a more recent version that
appears to successfully compile on Windows when using NEURON 8.2.7. This
commit may resolve #1052 but needs to be comprehensively tested.

* Reverse NEURON 8.2.6 cap

This reverses #1053 so the new `vecevent.mod`
file can be tested.

Note that 3.8 tests may still fail, due to NEURON 8.2.7
dropping support for Python 3.8.