[MRG] Add explicit base net regen (#1068)
* maint: add tiny script for recreating hier base

* doc: cleanup regen-base script and contrib

This makes some small cleanliness changes:

1. Removes the main usage and ruff-formats
`dev_scripts/regenerate_base_network.py`
2. Converts all code examples in `contributing.md` from RST to MD
format, and removes the `$` at the beginning of each line (which was
causing problems if you literally copy+pasted the command into zsh).
3. Adds a middle step in the Making your first PR section about when the
developer needs to have installed the dev version of HNN-Core.

* doc: expand and improve default net change guide

This adds better and *comprehensive* instructions to the Contributing
Guide on what to do if changes to the default network (and its files)
are necessary. As far as I know, this information is accurate to explain
which files to update, and *how* to update them.

* Add pre-regen backup to regen scripts

* merge regen make recipes, and update contrib