[MRG] doc: add Neymotin2020 impl to j09 docstring (#1056)
This is a tiny, trivial change to the doctring to the very important
function `network_models.py::jones_2009_model()`. This adds a note that
the "model network" is still defined by the Jones 2009 reference, but
that the "implementation" is from the Neymotin 2020 reference. This is
meant for rectifying the small but distinct differences between the two
references.

To be clear, although the code, parameters, and math between the the Jones
2009 implementation (can be seen https://modeldb.science/136803?tab=1 and
here https://github.com/ModelDBRepository/136803 ) and the Neymotin 2020
implementation (seen here https://github.com/jonescompneurolab/hnn ) are
very similar, they are not *exactly* the same.

In addition to differences with various parameters, there are also
differences between the MOD files. There are at least some differences
between the two implementations that affect the "speed" of activation
curves for some of the ion channels, which are documented in this
thread:
https://github.com/jonescompneurolab/hnn-core/discussions/930#discussioncomment-11230677

Therefore, adding this documentation change makes it clear to future
researchers that, even though Jones 2009 is the main paper of the model
in a scientific sense, they should refer to Neymotin 2020 when it comes
to the exact code and parameters used, especially for reproducibility.