[MRG] Chetan-Chunk-1: Introducing layer_dict for cell coordinates (#1095)
* rearranged

* z_coord cleanup

* testing layer dict

* WIP: changes to test_network.py

* ruff changes and cleaning out the code a bit

* maint: remove custom_positions arg

Because this will be done in a later Chetan-Chunk, we will remove it for now.

* doc: update docstring for _create_cell_coords

* maint: cleanup imports

* test: remove unnecessary default net test

The other `test_network_models` already basically tests the same
network, (i.e. `jones_2009_model` with its default params), so we don't
need to re-test this here. However, I have taken the `pos_dict`
size (equivalent to cell number) test and moved it into the other test
function since it's good.

* test: merge mesh shape tests into preexisting

There was already code to test custom mesh shapes. Some of the code in
`test_network_models_mod` that was applicable has been moved into the
preexisting `test_network_mesh` code. Similarly, the specific models
built with custom mesh shapes are now also tested for successful
simulation as well (and in parallel!).

* test: remove latter standard jones2009 test

This code was not really doing anything different than the standard
`jones_2009_model` construction code, which is already tested
elsewhere. Additionally, even if `n_pyr_x` etc. were changed, that is
still essentially testing different `mesh_shape` values, which is
already covered by `test_network_mesh`. Therefore, this testing code
is (AFAIK) unnecessary and removed.

* test: change name, abstract mesh shape

Change `test_network_models_mod` to a more specific name of
`test_custom_network_coords`. In the same test, this also abstracts the
shape of the custom network being created. This allows the test to be
parametrized to test against both a square and non-square mesh shape.

* doc: update Network docstring

* doc: add changelog entry for Chetan-Chunk1

---------

Co-authored-by: Austin E. Soplata <me@asoplata.com>