[MRG] test: remove linkchecks of specific github iss/PRs (#1069)
* test: remove linkchecks of specific github iss/PRs

This adds a new ignore regex to Sphinx's linkcheck, such that it will no
longer check the the high number of links that point to Github issues
and PRs in our documentation.

This group of links informs the vast majority of the links in our
documentation, and also the vast majority of the time that linkcheck
takes. These are also a frequent cause of linkcheck failures, due to
timeouts and other Github hosting issues that are not tied to errors in
our documentation linkage itself.

This also converts the former localhost ignore in linkcheck to one that
ignore references to localhost on all ports.

* doc/test: add optional, expanded "linkcheck-all"

This adds a new, optional type of linkcheck, called by using the
`doc/Makefile` recipe `linkcheck-all`. This runs linkcheck like the
regular linkcheck recipe, except that it ignores different links. This
is meant to be used whenever a developer wants to test that the links
now ignored in regular linkcheck by #1069 are still valid.

This new recipe only ignores links that:

1. follow any regular expressions used in `doc/conf.py`, or
2. are known to cause errors.

The links which are ignored currently are 3 paper DOI links (which
return 403 Forbidden errors if linkcheck'd), one reference to lines of
code in a specific commit of MNE which returns an error, and the
specific localhost URL for our GUI which also returns an error.

* upgrade linkcheck-all ignore to localhost regex

This improves the `doc/Makefile` `linkcheck-all` recipe to successfully
ignore all localhost URLs too.

* remove MNE anchor link from ignore due to #1086