maint: revert Dipole data expansion
It is too early to merge a change like this in, for at least the
following reasons:

1. There is still incomplete work (and bugs) preventing the current
state of this branch from successfully running a true "combined" dual
network simulation. There would need to be other work done before this
code is actually used, even for testing. That code is better suited to a
later PR instead of this one.

2. This only works for the case where there are exactly two celltypes
contributing to the dipole in each layer, instead of dynamically
organizing the data to fit multiple sizes.

3. Instead of hard-coding data-assignments in `dipole.py`, it would be
better to eventually (in a later PR) use the brand-new celltype
metadata (see here:
https://github.com/jonescompneurolab/hnn-core/blob/6b099a7dfe3f3ddbb3cd8c1472832089d80158bf/hnn_core/network_models.py#L84
) to organize the data shape.

Therefore, I will be removing expanded data shape code.