descope/revert: revert suffix/naming code
Originally, this example
script (`examples/howto/plot_dual_network_simulation.py`) made two
distinct networks which were different in two ways:

1. `net2`'s first GID was the first *after* the last GID of `net1` (with
or without `net1`'s drive), and

2. The celltypes in `net2` were renamed so that they all ended with a
new suffix, `_net2`.

These changes were made so that, *eventually*, in the future, when other
bugs were fixed and work was done, this same script would *then*
eventually be extended so that a user would simulate BOTH `net1` and
`net2` *simultaneously, as part of the same complete Network*. The
GID-reassignment code would be necessary to have the two Networks not
collide in GID space, and the celltype suffixes would be necessary to
differentiate between the appropriate celltype of `net1` versus `net2`.

However, after some investigation, I believe this PR needs to be
re-scoped to *only* include the GID-reassignment code, but *not* the
celltype-suffix code. The GID-reassignment code appears to work
currently, but the celltype-suffix does not.

Unfortunately, the celltype-suffix code is reliant on function
`Network._rename_cell_types` which exhibits the following:

1. It does NOT guarantee that dipole etc. output will be identical
"before vs. after" its use; see the following:
https://github.com/jonescompneurolab/hnn-core/blob/master/hnn_core/network.py#L1599
. It appears that, in some cases, it *was* producing identical output
before vs after in some prior commits of this PR. However, that may have
been a coincidence, since more work is needed to actually ensure that
`Network._rename_cell_types` is robust to Dipole calculation (including
testing that specifically, which would currently fail).

2. More importantly, the failure of `Network._rename_cell_types` to
produce dipole-identical output has been expanded by something in the
following recent (and large) commit:
https://github.com/jonescompneurolab/hnn-core/commit/6388f9f7c74f4943a1d9b06e9ae69314500913df
. In other words, Maira's dual-independent-networks scripts *before*
this commit do produce identical output in `net1` and `net2`. However,
after this commit, something has changed with either
`Network._rename_cell_types` or its dependencies, and the same
scripts *after* this commit do *not* produce identical output between
`net1` and `net2`.

As a result of this, I see two possible options to move forward in this
PR:

1. I fix `Network._rename_cell_types` to produce Dipole-identical
output, *including* while dealing with Chetan's new changes in the above
commit. Then, this PR includes both options for GID-reassignment and
that change. *I don't think this is a good idea, and would increase the
time-to-merge and scope significantly*.

2. I shrink the scope of this PR to exclude and remove any kind of
celltype-naming code, and make this PR exclusively about adding
GID-reassignment code. *I believe this is the best way forward*.

There was always going to be some clashing of Chetan's GSoC
main-objective of de-hardcoding celltypes, versus the minimum
changes (of the same sort) that Maira's GSoC objective needed in order
to lay the foundation and begin her work. Since both projects are still
incomplete, but there is *far* more work remaining to do for the
dual-network simulation feature, I think that changing this PR to *only*
be about adding GID-reassignment functionality is the best way forward.