[MRG] tiny: ignore all "venv" for codespell
This is a very tiny config change: this tells codespell to ignore all
directories and files matching the glob `*venv*`. This way, any local
virtual environment that includes the string "venv" such as at `my-venv`
or `.venv` will be ignored from codespell, which usually finds errors in
the files of the venvs.

This is useful in part because, *soon*, we will be able to have
full-feature HNN installs possible through venvs alone, without `conda`...