debug: works! looks TERRIBLE but works...
This suggests that the "Total Gain only updating on second receptor" bug
is tied to possible overwriting of duplicate `gain_text_input` or
`gain_indicator_output` objects. By forcefully making them unique (using
extremely lazy programming here, it's just for debug i swear!...), this
appears to fix the issue. Total gain texts are correctly updated, even
for connections multiple receptor types, AND those gains appear to be
accurately "deployed" when the behind-the-scenes Network object is
actually created for a sim.

Needs cleanup and much further testing, but the hardest part of this
feature may be over.