feat: add warning about non-uniform syn gains
This adds a new private checker function that actually checks if the
"uniform" assumption is held within connections of each "global" synapse
gain type. If the assumption doesn't hold, then a warning is printed,
either to the log in the case of the GUI (due to its use of
`Network.get_global_synaptic_gains`, which now includes the check) or to
stdout in the case of the API (due to addition of the check into
`hnn_io.read_network_configuration`).

This also fixes a small BoundedFloatText issue when loading Networks
containing non-uniform synaptic gains.