feat: Part 1 of many of extracting KD's v_init
This is the first commit of many to come where I'm going to extract the
parts of #1168 relevant to the new, proper voltage initialization.

The plan is:
1. We'll separate the changes to the "v_init" stuff and make it its own
PR onto `master`,
2. Change the tests that become broken as we go along,
3. Everyone reviews the changes (which will likely end up in a slightly
different form from how they currently exist in #1168),
4. We merge these voltage initialization changes into `master`,
5. Then finally, we merge/rebase the voltage initialization
changes (which are now on `master`) *back* into the code at #1168 .
6. Repeat the process for all code features included in the length #1168
work, until everything gets merged, piece-by-piece.

For starters, this commit consists of *only* the extracted changes to
`cell.py` and nothing else. Fortunately, in this case, the tests seem to
pass without having to change anything.

The original person who actually wrote this code was @katduecker, which
is why they are indicated as author. This version of the code that was
extracted from #1168 was done so by @asoplata .