fix: add missing "v" read-in (with back-compat)
This also includes regenerated networks, since doing that is necessary
to allow the tests to pass as well.

The original person who actually wrote this code was @katduecker, which
is why they are indicated as author. This version of the code that was
extracted from #1168 was done so by @asoplata .