fix: add missing L2 soma v_init, remove old init
This fixes what I think is a bug currently present in #1168, where the
L2's `soma` section did not have its universal voltage-initialization
value applied (but the dendrites did).

Additionally, @katduecker , I think the same bug is present in your
`_get_interneuron_soma()` function (which I've removed in this commit
since it's not relevant to the voltage-initialization). Your
`_get_interneuron_soma` takes a `v_init` value, but does not apply it in
`Section` creation. This was probably missed.

This also removes `NetworkBuilder.state_init()` like KD did, since it is
no longer needed, as the voltage initialization is completely moved over
to `cells_default.py` now.

With this, I think the voltage-initialization stuff has been completely
extracted from #1168. The only remaining work is to write new tests
specifically for the `v_init` where it is used.