feat: add KD's new Ca-baseline-correction
This extracts the code needed to deploy @katduecker's
`Dipole._baseline_renormalize_ca` to new simulations, including addition
of the necessary Network file used for it.

This does not include any tests for the new renorm function, but ideally
that is best to be done after some Dipole data is generated using it
with which to compare it to.

Authorship set to Katharina since she actually wrote the code