feat: KD add remaining extracted non-model code
I believe this adds the last of the code that is extracted from #1168
that is *not* the actual Duecker model itself, but instead code changes
that are used by the default model. These functions are not currently
tested at all, and I have not added tests for them for similar reasons
to
https://github.com/jonescompneurolab/hnn-core/pull/1185#issuecomment-3502924088

After this, we should be able to make a PR that adds the actual Duecker
model itself. However, there will be additional work for that step: we
need to update `hnn_io.read_network_configuration` and the equivalent
`...write...` such that they support the Duecker model.

Continuing the pattern, this is based on top of #1187.