patch: add possible fix for Duecker file read-in
Hey @katduecker, this *may* be all that is necessary to enable reading
and writing of Network files (via the "hierarchical JSON" version) of
your new Duecker model. I've had to add in a patch from
https://github.com/jonescompneurolab/hnn-core/pull/1185/commits/43c77c1ac945771228346bda87d7f7559c2ef865
in order to get initial voltage loaded properly, but the other changes
in `hnn_io.py` seem to be all that is required in my limited, local
testing.

Can you please test that this enables successful reading and writing of
your new Duecker model into Network files?