fix: target opt data RMSE widget resets correctly
This fixes a small bug where, before, if there was a pre-existing
simulation data set present (such as from an earlier regular run or
`maximize_psd` optimization run), then the `dipole_rmse` "Target Data"
widget would incorrectly use the first value of its list as the target
data, instead of the selected value. This turned out to be because, in
the changed code, setting the options of the widget using

```
self._external_data_widget.options = all_sim_names
```

actually changes the value to the first entry in the list, even if the
prior value is a legitimate element of the new `options`. This commit
adds a tiny amount of code to workaround this, so that the old value is
retained, and the user-selected option for Target Data is used, even in
the event of there being a pre-existing simulation that is indexed
before any loaded data.