maint: read in pos_dict and cell_types
This is related to #1168, and is required in order for @katduecker's new
model to be successfully read in from Network JSON files.

This makes a very tiny change in `hnn_io::dict_to_network`, affecting
`hnn_io::read_network_configuration` among others. When a dictionary
of `Network` parameters is being read in (such as from file), the
`Network.pos_dict` and `Network.cell_types` data is passed from the
originating dictionary to the `Network` initializer, instead of being
added to the `Network` object after creation. This takes advantage of
Chetan's updates to `Network` that can use these new arguments. This
allows Network JSON files with custom `cell_types` and/or `pos_dict` to
have that information read in, instead of having it overwritten with the
defaults, which is what is done currently.